[Refactor] Split connector into several files
