+++ be/src/connector/CMakeLists.txt
@@ -16,6 +16,9 @@ set(LIBRARY_OUTPUT_PATH "${BUILD_DIR}/src/connector")
 
 ADD_BE_LIB(Connector
         connector.cpp
+        data_source.cpp
+        connector_registry.cpp
+        connector_registry_init.cpp
         hive_connector.cpp
         es_connector.cpp
         jdbc_connector.cpp
