+++ be/src/connector/file_chunk_sink.h
@@ -21,11 +21,10 @@
 
 #include "common/status.h"
 #include "common/thread/priority_thread_pool.hpp"
-#include "connector/connector.h"
-#include "connector_chunk_sink.h"
+#include "connector/connector_chunk_sink.h"
+#include "connector/utils.h"
 #include "formats/column_evaluator.h"
 #include "formats/file_writer.h"
-#include "utils.h"
 
 namespace starrocks::connector {
